Revert "Evaluation file Code Restructure"
